requirements.txt
flask-2.0.2
Jinja2-3.0.3
MarkupSafe-2.0.1
Werkzeug-2.0.2
click-8.0.3
itsdangerous-2.0.1
